backend/requirements.txt
django
djangorestframework
pandas
matplotlib
PyQt5
django-cors-headers
requests
gunicorn
whitenoise
